EfiRom: allow a DeviceId value of 0
We are using the EfiRom tool to convert the Intel GOP driver to an EFI
ROM. This ROM is added to the luemuctl repository and the VendorId
and DeviceId are patched to valid values right before starting a
bhyve VM. The ROM added to the luemuctl repository should contain
an invalid value. This ensures that our CI test will fail if
luemuctl can't patch the ROM. Therefore, allow the invalid value
of 0 but log an error to inform users about it.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>